refactor(modules): extract logic for two browser fixes into two modules
