+++ src/utils/iosFix.js
@@ -1,0 +1,21 @@
+import * as dom from '../utils/dom'
+import { swalClasses } from '../utils/classes'
+
+// Fix iOS scrolling http://stackoverflow.com/q/39626302/1331425
+export const iOSfix = () => {
+  const iOS = /iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream
+  if (iOS && !dom.hasClass(document.body, swalClasses.iosfix)) {
+    const offset = document.body.scrollTop
+    document.body.style.top = (offset * -1) + 'px'
+    dom.addClass(document.body, swalClasses.iosfix)
+  }
+}
+
+export const undoIOSfix = () => {
+  if (dom.hasClass(document.body, swalClasses.iosfix)) {
+    const offset = parseInt(document.body.style.top, 10)
+    dom.removeClass(document.body, swalClasses.iosfix)
+    document.body.style.top = ''
+    document.body.scrollTop = (offset * -1)
+  }
+}
