+++ src/utils/scrollbarFix.js
@@ -1,0 +1,21 @@
+import * as dom from '../utils/dom'
+
+export const fixScrollbar = () => {
+  // for queues, do not do this more than once
+  if (dom.states.previousBodyPadding !== null) {
+    return
+  }
+  // if the body has overflow
+  if (document.body.scrollHeight > window.innerHeight) {
+    // add padding so the content doesn't shift after removal of scrollbar
+    dom.states.previousBodyPadding = document.body.style.paddingRight
+    document.body.style.paddingRight = dom.measureScrollbar() + 'px'
+  }
+}
+
+export const undoScrollbar = () => {
+  if (dom.states.previousBodyPadding !== null) {
+    document.body.style.paddingRight = dom.states.previousBodyPadding
+    dom.states.previousBodyPadding = null
+  }
+}
